corrected integration test,refactored code,added comments,added exception classes

+++ itv/src/main/java/com/itv/dto/ItemDetailDto.java
@@ -1,15 +1,19 @@
 package com.itv.dto;
 
+/**
+ * @author Likin Gera
+ *
+ */
 public class ItemDetailDto {
 	
-	private int id;
+	private Long id;
 	private String itemName;
 	private int unitPrice;
 	
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getItemName() {
